+++ src/d-table/utils/filter.js
@@ -6,7 +6,7 @@ export const search = (
 ) => {
   return rows.filter((row) => {
     for (let key of Object.keys(row)) {
-      if (columnHashMap[key].filterable) {
+      if (columnHashMap[key] && columnHashMap[key].filterable) {
         let hayStack = caseSensitiveSearch ? row[key] : row[key].toLowerCase();
         let needle = caseSensitiveSearch ? search : search.toLowerCase();
         if (hayStack.includes(needle)) {